refactor pagination for data-table-demo
