--- apps/devfractal-demo/src/table-lib/table-pagination/types.ts
@@ -1,14 +1,0 @@
-export type PaginationProps = {
-  count: number
-  currentPage: number
-  onNext: () => void
-  onPrev: () => void
-}
-
-export type PaginationProps2 = {
-  count: number
-  currentPage: number
-  onNext: () => void
-  onPrev: () => void
-  onClick: (page: number) => void
-}
